added commit to model name and logs
